refactor and remove duplicate code
